feat(teams): add ibm platform metrics in secure teams
